completed part.. atlast.. Yaaay

+++ part4/BlogList/controllers/blogs.js
@@ -1,14 +1,17 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blogs')
+const User = require('../models/users');
+const jwt = require('jsonwebtoken')
 
-
-
-blogRouter.get('/', (request, response) => {
-  Blog
+blogRouter.get('/', async (req, res) => {
+  const blogs = await Blog
     .find({})
-    .then(blogs => {
-      response.json(blogs)
+    .populate('User', {
+      username: 1,
+      name: 1
     })
+
+  res.json(blogs) //.map(n=> n.toJSON())
 })
 
 
@@ -24,26 +27,62 @@ blogRouter.get('/:id', async (request, response, next) => {
     }
   }).catch(e => next(e))
 
-
 })
 
-blogRouter.post('/', (request, response) => {
-  const blog = new Blog(request.body)
-  if (!blog.hasOwnProperty('likes')) {
-    blog['likes'] = 0;
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
+
+blogRouter.post('/', async (request, response) => {
+  const body = new Blog(request.body)
+
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  if (!request.body.hasOwnProperty('likes')) {
+    request.body['likes'] = 0;
   }
   if (!request.body.hasOwnProperty('title') || !request.body.hasOwnProperty('url')) {
     console.log('missing props');
-    return response.status(400).json({ error: 'content missing' })
-  }
-  blog
-    .save()
-    .then(result => {
-      response.status(201).json(result)
+    return response.status(400).json({
+      error: 'content missing'
     })
+  }
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+
+  })
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.json(savedBlog)
+
 })
 
 blogRouter.delete('/:id', async (req, res) => {
+
+  const blog = await Blog.findById(req.params.id)
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+  if(blog.user.toString() !== decodedToken.id.toString()){
+    return res.status(401).json({error:'permission denied!'})
+  }
+
   await Blog.findByIdAndRemove(req.params.id)
   res.status(204).end()
 })
@@ -57,7 +96,9 @@ blogRouter.put('/:id', (req, res, next) => {
   }
 
   console.log('blog ', blog)
-  Blog.findByIdAndUpdate(req.params.id, blog, { new: true }).then(updatedBlog => {
+  Blog.findByIdAndUpdate(req.params.id, blog, {
+    new: true
+  }).then(updatedBlog => {
     res.json(updatedBlog.toJSON())
   }).catch(e => next(e))
 })
